+++ components/glow.lua
@@ -3,7 +3,7 @@ local AddonName, SAO = ...
 -- List of known ActionButton instances that currently match one of the spell IDs to track
 -- This does not mean that buttons are glowing right now, but they could glow at any time
 -- key = glowID (= spellID of action), value = list of ActionButton objects for this spell
--- (side note: the sublist of buttons is a map of key = action slot and value = button)
+-- (side note: the sublist of buttons is a table of key = button and a dummy value = true)
 -- The list will change each time an action button changes, which may happen very often
 -- For example, any macro with [mod:shift] updates the list every time Shift is pressed
 SAO.ActionButtons = {}
@@ -50,14 +50,8 @@ end
 -- If forceRefresh is true, refresh even if old spell ID and new spell ID are identical
 -- Set forceRefresh if the spell ID of the button may switch from untracked to tracked (or vice versa) in light of recent events
 function SAO.UpdateActionButton(self, button, forceRefresh)
-    local oldAction = button.lastAction; -- Set by us, a few lines below
     local oldGlowID = button.lastGlowID; -- Set by us, a few lines below
-    local newAction = button.action; -- Set by the game, inside Blizzard's ActionButton.lua
-    local newGlowID = nil;
-    if HasAction(button.action) then
-        newGlowID = self:GetSpellIDByActionSlot(button.action);
-    end
-    button.lastAction = newAction; -- Write button.lastAction here, but use oldAction/newAction for the rest of the function
+    local newGlowID = button:GetGlowID();
     button.lastGlowID = newGlowID; -- Write button.lastGlowID here, but use oldGlowID/newGlowID for the rest of the function
 
     if (oldGlowID == newGlowID and not forceRefresh) then
@@ -67,17 +61,17 @@ function SAO.UpdateActionButton(self, button, forceRefresh)
 
     -- Register/unregister button as 'dormant' i.e., not tracked but could be tracked in the future
     if (oldGlowID and not self.RegisteredGlowSpellIDs[oldGlowID] and type(self.DormantActionButtons[oldGlowID]) == 'table') then
-        if (self.DormantActionButtons[oldGlowID][oldAction] == button) then
-            self.DormantActionButtons[oldGlowID][oldAction] = nil;
+        if (self.DormantActionButtons[oldGlowID][button] == button) then
+            self.DormantActionButtons[oldGlowID][button] = nil;
         end
     end
     if (newGlowID and not self.RegisteredGlowSpellIDs[newGlowID]) then
         if (type(self.DormantActionButtons[newGlowID]) == 'table') then
-            if (self.DormantActionButtons[newGlowID][newAction] ~= button) then
-                self.DormantActionButtons[newGlowID][newAction] = button;
+            if (self.DormantActionButtons[newGlowID][button] ~= button) then
+                self.DormantActionButtons[newGlowID][button] = button;
             end
         else
-            self.DormantActionButtons[newGlowID] = { [newAction] = button };
+            self.DormantActionButtons[newGlowID] = { [button] = button };
         end
     end
 
@@ -89,23 +83,23 @@ function SAO.UpdateActionButton(self, button, forceRefresh)
     -- Untrack previous action button and track the new one
     if (oldGlowID and self.RegisteredGlowSpellIDs[oldGlowID] and type(self.ActionButtons[oldGlowID]) == 'table') then
         -- Detach action button from the former glow ID
-        if (self.ActionButtons[oldGlowID][oldAction] == button) then
-            self.ActionButtons[oldGlowID][oldAction] = nil;
+        if (self.ActionButtons[oldGlowID][button] == button) then
+            self.ActionButtons[oldGlowID][button] = nil;
         end
     end
     if (newGlowID and self.RegisteredGlowSpellIDs[newGlowID]) then
         if (type(self.ActionButtons[newGlowID]) == 'table') then
             -- Attach action button to the current glow ID
-            if (self.ActionButtons[newGlowID][newAction] ~= button) then
-                self.ActionButtons[newGlowID][newAction] = button;
+            if (self.ActionButtons[newGlowID][button] ~= button) then
+                self.ActionButtons[newGlowID][button] = button;
             end
         else
             -- This glow ID has no Action Buttons yet: be the first
-            self.ActionButtons[newGlowID] = { [newAction] = button };
+            self.ActionButtons[newGlowID] = { [button] = button };
         end
         -- Remove from the 'dormant' table, if it was dormant
-        if (type(self.DormantActionButtons[newGlowID]) == 'table' and self.DormantActionButtons[newGlowID][newAction] == button) then
-            self.DormantActionButtons[newGlowID][newAction] = nil;
+        if (type(self.DormantActionButtons[newGlowID]) == 'table' and self.DormantActionButtons[newGlowID][button] == button) then
+            self.DormantActionButtons[newGlowID][button] = nil;
         end
     end
 
@@ -114,22 +108,39 @@ function SAO.UpdateActionButton(self, button, forceRefresh)
     local mustGlow = newGlowID and (self.GlowingSpells[newGlowID] ~= nil);
 
     if (not wasGlowing and mustGlow) then
-        if (not SpellActivationOverlayDB.glow or SpellActivationOverlayDB.glow.enabled) then
-            ActionButton_ShowOverlayGlow(button);
+        if (not SpellActivationOverlayDB or not SpellActivationOverlayDB.glow or SpellActivationOverlayDB.glow.enabled) then
+            button:EnableGlow();
         end
     elseif (wasGlowing and not mustGlow) then
-        ActionButton_HideOverlayGlow(button);
+        button:DisableGlow();
     end
 end
 
 -- Grab all action button activity that allows us to know which button has which spell
-function HookActionButton_Update(self)
-    if (self:GetParent() == OverrideActionBar) then
+function HookActionButton_Update(button)
+    if (button:GetParent() == OverrideActionBar) then
         -- Act on all buttons but the ones from OverrideActionBar, whatever that is
         return;
     end
 
-    SAO:UpdateActionButton(self);
+    if (not button.GetGlowID) then
+        button.GetGlowID = function(button)
+            if (button.action and HasAction(button.action)) then
+                return SAO:GetSpellIDByActionSlot(button.action);
+            end
+        end
+    end
+    if (not button.EnableGlow) then
+        button.EnableGlow = function(button)
+            ActionButton_ShowOverlayGlow(button);
+        end
+    end
+    if (not button.DisableGlow) then
+        button.DisableGlow = function(button)
+            ActionButton_HideOverlayGlow(button);
+        end
+    end
+    SAO:UpdateActionButton(button);
 end
 hooksecurefunc("ActionButton_Update", HookActionButton_Update);
 
@@ -162,8 +173,8 @@ function SAO.AddGlowNumber(self, spellID, glowID)
     else
         self.GlowingSpells[glowID] = { [spellID] = true };
         for _, frame in pairs(actionButtons or {}) do
-            if (not SpellActivationOverlayDB.glow or SpellActivationOverlayDB.glow.enabled) then
-                ActionButton_ShowOverlayGlow(frame);
+            if (not SpellActivationOverlayDB or not SpellActivationOverlayDB.glow or SpellActivationOverlayDB.glow.enabled) then
+                frame:EnableGlow();
             end
         end
     end
@@ -218,8 +229,60 @@ function SAO.RemoveGlow(self, spellID)
             self.GlowingSpells[glowSpellID] = nil;
             local actionButtons = self.ActionButtons[glowSpellID];
             for _, frame in pairs(actionButtons or {}) do
-                ActionButton_HideOverlayGlow(frame);
+                frame:DisableGlow();
             end
         end
     end
 end
+
+-- Track PLAYER_LOGIN which happens immediately after all ADDON_LOADED events
+-- Which means, at this point we know which addons are installed and loaded
+local binder = CreateFrame("Frame", "SpellActivationOverlayLABBinder");
+binder:RegisterEvent("PLAYER_LOGIN");
+binder:SetScript("OnEvent", function()
+    local LAB = LibStub("LibActionButton-1.0", true);
+    local LAB_ElvUI = LibStub("LibActionButton-1.0-ElvUI", true);
+    local LBG = LibStub("LibButtonGlow-1.0", true);
+    local LCG = LibStub("LibCustomGlow-1.0", true);
+
+    local buttonUpdateFunc = function(libGlow, event, self)
+        if (not self.GetGlowID) then
+            self.GetGlowID = self.GetSpellId;
+        end
+        if (not self.EnableGlow) then
+            self.EnableGlow = function(button)
+                libGlow.ShowOverlayGlow(button);
+            end
+        end
+        if (not self.DisableGlow) then
+            self.DisableGlow = function(button)
+                libGlow.HideOverlayGlow(button);
+            end
+        end
+        SAO:UpdateActionButton(self);
+    end
+
+    local LBGButtonUpdateFunc = function(event, self)
+        buttonUpdateFunc(LBG, event, self);
+    end
+
+    local LCGButtonUpdateFunc = function(event, self)
+        buttonUpdateFunc(LCG, event, self);
+    end
+
+    -- Support for LibActionButton e.g., used by Bartender
+    if (LAB and LBG) then -- Prioritize LibButtonGlow, if available
+        LAB:RegisterCallback("OnButtonUpdate", LBGButtonUpdateFunc);
+    elseif (LAB and LCG) then -- Otherwise fall back to LibCustomGlow
+        LAB:RegisterCallback("OnButtonUpdate", LCGButtonUpdateFunc);
+    end
+
+    -- Support for ElvUI's LibActionButton
+    if (LAB_ElvUI and LCG) then -- Prioritize LibCustomGlow, if available
+        LAB_ElvUI:RegisterCallback("OnButtonUpdate", LCGButtonUpdateFunc);
+    elseif (LAB_ElvUI and LBG) then -- Otherwise fall back to LibButtonGlow
+        LAB_ElvUI:RegisterCallback("OnButtonUpdate", LBGButtonUpdateFunc);
+    end
+
+    binder:UnregisterEvent("PLAYER_LOGIN");
+end);